refactor: move psbt prelude.rs to macros.rs
